requirements.txt
requests
beautifulsoup4
openai
anthropic
python-dotenv
